Add proper ARRAY<> and MAP<> translations to dbDataType() and always translate POSIXct to TIMESTAMP to maximize compatibility

+++ tests/testthat/test-response.to.content.R
@@ -45,9 +45,9 @@ with_locale(test.locale(), test_that)(".response.to.content works", {
         ),
         list(
           name = "m",
-          type = "bigint",
+          type = "integer",
           typeSignature = list(
-            rawType = "bigint",
+            rawType = "integer",
             typeArguments = list(),
             literalArguments = list()
           )
